requirements.txt
gensim==4.0.0
nltk==3.6.2
numpy==1.19.5
pandas==1.2.4
pathlib2==2.3.5
scikit-learn==0.24.2
scipy==1.6.3
segeval==2.0.11
sentence-transformers==1.1.0
tensorflow==2.2.0
tensorflow-hub==0.12.0
termcolor==1.1.0
torch==1.9.0
tqdm==4.60.0
transformers==4.5.1
umap-learn==0.5.1
xlrd==2.0.1
XlsxWriter==1.4.2
pytorch-lightning==1.3.1